change: deleted shap_wrappers_combined and moved code to shap_wrappers
